Share the internal-reference integrity guard across both parsers
`stripInvalidArticleLinks` and `enforceInternalReferenceIntegrity` were
duplicated verbatim in fmxParser.mjs and eurlex-html-parser.js — exactly
the kind of cross-format parsing mechanic legal-reference-core.cjs exists
to own, and a drift risk (a fix to one wouldn't reach the other).

Move both into legal-reference-core.cjs and import them:
- eurlex-html-parser.js: drop both local copies, import the shared pair.
- fmxParser.mjs: import both; the final integrity-guard block becomes a
  single enforceInternalReferenceIntegrity({ articles, recitals, annexes,
  crossReferences }) call. The shared guard also covers annex HTML (fmx
  previously stripped it inline; still does, and the guard is idempotent)
  and defensively covers article.paragraphs[].html.

Behavior-preserving: output shape is unchanged, so no PARSER_VERSION bump.
Verified: legal-reference-core + eurlex-html-parser + case-law-parser
node:test suites pass, fmxParser vitest suite passes (110), full backend
`npm test` green (249 pass / 2 skipped).
